HUD and Menus
Added player HUD to the game scene, didn't code the health/mana bars or  skill key shaders - cuz not sure if there was a plan already.

All the menus are ready, attached the buttons to the proper scripts and they work. Added in the settings menu except only one option is on there (not coded yet) for the future to finish up/work with. idk what the other two options were gonna be.

Added in the canvas for win/lose states.... wasn't sure if you wanted them as scenes or not, couldn't remember how we did it last time